pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anything2anki"
version = "0.1.0"
description = "Generate Anki cards from text files using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aisuite[openai]>=0.1.12",
    "genanki>=0.13.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
anything2anki = "anything2anki.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/anything2anki"]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
